+++ src/commonTest/kotlin/com/github/ajalt/colormath/LABTest.kt
@@ -1,15 +1,17 @@
 package com.github.ajalt.colormath
 
-import io.kotlintest.data.forall
-import io.kotlintest.matchers.doubles.plusOrMinus
-import io.kotlintest.shouldBe
-import io.kotlintest.tables.row
-import org.junit.Test
+import io.kotest.data.blocking.forAll
+import io.kotest.data.row
+import io.kotest.matchers.doubles.plusOrMinus
+import io.kotest.matchers.shouldBe
+import kotlin.js.JsName
+import kotlin.test.Test
 
 class LABTest {
     @Test
+    @JsName("LAB_to_XYZ")
     fun `LAB to XYZ`() {
-        forall(
+        forAll(
                 row(LAB(000.0, 000.0, 000.0), 0.0, 0.0, 0.0),
                 row(LAB(050.0, 050.0, 050.0), 028.454, 018.419, 003.533),
                 row(LAB(075.0, 075.0, 075.0), 077.563, 048.278, 007.476),